Inlines Options class.
