+++ lib/nutrella.rb
@@ -3,7 +3,6 @@ require "active_support/core_ext/string"
 require "nutrella/cache"
 require "nutrella/command"
 require "nutrella/configuration"
-require "nutrella/options"
 require "nutrella/task_board"
 require "nutrella/version"
