+++ lib/nutrella/command.rb
@@ -1,3 +1,6 @@
+require "git"
+require "optparse"
+
 module Nutrella
   #
   # Knows how to dispatch to a command based on the specified options.
@@ -5,61 +8,43 @@ module Nutrella
   class Command
     attr_reader :options
 
-    def initialize(options)
-      @options = options
+    def initialize(args)
+      @args = args
     end
 
     def task_board
-      options.parse
-      dispatch
+      boardname = nil
+
+      OptionParser.new do |opts|
+        opts.on("-g", "--current-git-branch", "Open the board matching the current git branch")
+        opts.on("-t", "--trello-board BOARD", "Open the board with name BOARD") { |n| boardname = n }
+        opts.on("--init", "Initialize the nutrella.yml configuration") { |n| Configuration.new.write_default; exit }
+        opts.on("-v", "--version", "Display the version") { |n| puts Nutrella::VERSION; exit }
+        opts.on("-h", "--help", "Display this screen") { |n| puts opts; exit }
+      end.parse!(@args)
+
+      find_or_create_board(boardname || trello_board_name_derived_from_git_branch)
     rescue => e
-      puts "Error: #{e}"
-      exit(false)
+      abort "Error: invalid option: #{@args}"
     end
 
     private
 
-    def dispatch
-      case
-      when options.init?
-        write_default_configuration
-      when options.show_usage?
-        puts options.usage
-      when options.show_version?
-        puts options.version
-      else
-        task_board_url
-      end
-    end
+    def find_or_create_board(board_name)
+      task_board = TaskBoard.new(board_name, Configuration.new)
+      board = task_board.find
+      return board if board
 
-    def write_default_configuration
-      Configuration.new.write_default
-    end
-
-    def task_board_url
-      (cached_url || find_or_create_board).tap do |board|
-        Cache.write(options.board_name, board.url) unless board.nil?
-      end
-    end
-
-    def cached_url
-      url = Cache.lookup(options.board_name)
-
-      url && OpenStruct.new(url: url)
-    end
-
-    def find_or_create_board
-      task_board = TaskBoard.new(options.board_name, Configuration.new)
-
-      existing = task_board.find
-      return existing unless existing.nil?
-
-      task_board.create if confirm_create? task_board
+      task_board.create if confirm_create?(task_board)
     end
 
     def confirm_create?(task_board)
       print "Create the '#{task_board.name}' task board? [y/N]: "
       gets =~ /^y/i
+    end
+
+    def trello_board_name_derived_from_git_branch
+      Git.open(".").current_branch.humanize.titleize
     end
   end
 end
